Fix parameters type in methods from int to long

+++ Store.BusinessLogic/Services/Interfaces/IUserService.cs
@@ -6,10 +6,10 @@ namespace Store.BusinessLogic.Services.Interfaces
     public interface IUserService
     {
         IEnumerable<ApplicationUser> GetUsers();
-        ApplicationUser GetUserById(int userId);
+        ApplicationUser GetUserById(long userId);
         void AddUser(ApplicationUser user);
         void UpdateUser(ApplicationUser user);
-        void DeleteUser(int userId);
+        void DeleteUser(long userId);
         void Save();
     }
 }
